requirements.txt
streamlit
torch
torchvision
timm
pillow
numpy
opencv-python
grad-cam
gdown

